+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [8.0.1] (IN PROGRESS)
 
 * Stop infinity loading of package titles when a user edits the package record while they are being added/removed to holdings. (UIEH-1349)
+* Only display tags that have already been added to records as options when the user tries to search for records. (UIEH-1357)
 
 ## [8.0.0] (https://github.com/folio-org/ui-eholdings/tree/v8.0.0) (2023-02-15)